+++ platform/ui/src/elements/form/Select.js
@@ -7,8 +7,18 @@ import PropTypes from 'prop-types';
 class Select extends Component {
   constructor(props) {
     super(props);
-    this.state = { value: this.props.value };
   }
+
+  static propTypes = {
+    options: PropTypes.arrayOf(
+      PropTypes.shape({
+        key: PropTypes.string.isRequired,
+        value: PropTypes.string.isRequired,
+      })
+    ),
+    value: PropTypes.string,
+    onChange: PropTypes.func,
+  };
 
   handleChange = event => {
     const value = event.target.value;
@@ -18,32 +28,22 @@ class Select extends Component {
 
   render() {
     return (
-      <select
-        className="select-ohif"
-        value={this.state.selected}
-        onChange={this.handleChange}
-      >
-        {this.props.options.map(({ key, value }) => {
-          return (
-            <option key={key} value={value}>
-              {key}
-            </option>
-          );
-        })}
-      </select>
+      <div className="select-ohif-container">
+        <label className="select-ohif-label" htmlFor={this.id}>
+          {this.props.label}
+          <select className="form-control select-ohif" {...this.props}>
+            {this.props.options.map(({ key, value }) => {
+              return (
+                <option key={key} value={value}>
+                  {key}
+                </option>
+              );
+            })}
+          </select>
+        </label>
+      </div>
     );
   }
 }
-
-Select.propTypes = {
-  options: PropTypes.arrayOf(
-    PropTypes.shape({
-      key: PropTypes.string.isRequired,
-      value: PropTypes.string.isRequired,
-    })
-  ),
-  value: PropTypes.string,
-  onChange: PropTypes.func,
-};
 
 export { Select };