+++ platform/viewer/cypress/integration/pwa/OHIFExtensionPDF.spec.js
@@ -20,7 +20,8 @@ describe('OHIF PDF Extension', () => {
       .its('length')
       .should('be.eq', 1);
 
-    cy.screenshot();
-    cy.percySnapshot();
+    // This won't work unless we switch to an extension that renders using `canvas`
+    // Currently, we rely on the browser's built-in implementation
+    cy.percyCanvasSnapshot('PDF Extension');
   });
 });