+++ platform/viewer/src/components/OHIFLogo/OHIFLogo.js
@@ -12,6 +12,13 @@ function OHIFLogo() {
       href="http://ohif.org"
     >
       <Icon name="ohif-logo" className="header-logo-image" />
+      {/* Logo text would fit smaller displays at two lines:
+       *
+       * Open Health
+       * Imaging Foundation
+       *
+       * Or as `OHIF` on really small displays
+       */}
       <Icon name="ohif-text-logo" className="header-logo-text" />
     </a>
   );